Extract getProxyAccount
